Made a C++ package to randomly distribute heights
Takes the slam_cloud topic and creates a new pointcloud with additional Z value randomly distributed between 0 and 1 (float). Working on integrating the IMU from the pixhawk to handle pitch, will see if it attempts to factor vertical acceleration also. Then will integrate with the TFMini to get the current height. Need to get DIY sensor working with this also.